Add TryCatchNode to represent try-catch operation (#50)
* Add TryCatchNode to represent try-catch operation

Fixes #14

+++ src/NodeDev.Tests/GraphExecutorTests.cs
@@ -29,7 +29,6 @@ public class GraphExecutorTests
 		entryNode = new EntryNode(graph);
 
 		returnNode = new ReturnNode(graph);
-		returnNode.Inputs.Add(new("Result", entryNode, nodeClass.TypeFactory.Get<TOut>()));
 
 		addNode = new Core.Nodes.Math.Add(graph);
 
@@ -190,7 +189,8 @@ public class GraphExecutorTests
 		finally
 		{
 			// Clean up
-			Directory.Delete(buildOptions.OutputPath, true);
+			if(Directory.Exists(buildOptions.OutputPath))
+				Directory.Delete(buildOptions.OutputPath, true);
 		}
 	}
 
@@ -235,7 +235,6 @@ public class GraphExecutorTests
 		graph.Connect(addNode.Outputs[0], smallerThan.Inputs[0], false);
 
 		var returnNode2 = new Core.Nodes.Flow.ReturnNode(graph);
-		returnNode2.Inputs.Add(new("Result", entryNode, graph.SelfClass.TypeFactory.Get<int>()));
 		returnNode2.Inputs[1].UpdateTextboxText("0");
 		graph.AddNode(returnNode2, false);
 
@@ -272,7 +271,6 @@ public class GraphExecutorTests
 
 		var returnNode2 = new Core.Nodes.Flow.ReturnNode(graph);
 		graph.AddNode(returnNode2, false);
-		returnNode2.Inputs.Add(new("Result", entryNode, graph.SelfClass.TypeFactory.Get<int>()));
 		returnNode2.Inputs[1].UpdateTextboxText("0");
 
 		var branchNode = new Core.Nodes.Flow.Branch(graph);
@@ -289,5 +287,161 @@ public class GraphExecutorTests
 
 		output = Run<int>(graph.Project, options, [-1, -2]);
 		Assert.Equal(1, output);
+	}
+
+    // This test validates the TryCatchNode by simulating a scenario where an exception is thrown and caught.
+    // The test sets up a graph with an entry node, a TryCatchNode, and return nodes for both try and catch blocks.
+    // The try block attempts to parse an invalid integer string, which throws an exception.
+    // The catch block returns 1, indicating that the exception was caught.
+    // The test asserts that the output is 1, confirming that the exception was caught and handled correctly.
+    [Theory]
+    [MemberData(nameof(GetBuildOptions))]
+    public void TestTryCatchNode(SerializableBuildOptions options)
+    {
+        var project = new Project(Guid.NewGuid());
+        var nodeClass = new NodeClass("Program", "Test", project);
+        project.Classes.Add(nodeClass);
+
+        var graph = new Graph();
+        var method = new NodeClassMethod(nodeClass, "MainInternal", nodeClass.TypeFactory.Get<int>(), graph);
+        method.IsStatic = true;
+        nodeClass.Methods.Add(method);
+        graph.SelfMethod = nodeClass.Methods.First();
+
+        var entryNode = new EntryNode(graph);
+        var tryCatchNode = new TryCatchNode(graph);
+		tryCatchNode.Outputs[3].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<Exception>(), overrideInitialType: true);
+
+        graph.AddNode(entryNode, false);
+        graph.AddNode(tryCatchNode, false);
+
+		// Create local variable
+		var declareVariableNode = new DeclareVariableNode(graph);
+		declareVariableNode.Inputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Initial value
+		declareVariableNode.Outputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Variable
+		graph.AddNode(declareVariableNode, false);
+
+		graph.Connect(entryNode.Outputs[0], declareVariableNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[0], tryCatchNode.Inputs[0], false);
+
+		var returnNode = new ReturnNode(graph);
+		graph.Connect(declareVariableNode.Outputs[1], returnNode.Inputs[1], false);
+		graph.AddNode(returnNode, false);
+
+		// Create the catch block body
+		var catchVariableNode = new SetVariableValueNode(graph);
+		catchVariableNode.Inputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Variable
+		catchVariableNode.Inputs[2].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Value
+		catchVariableNode.Inputs[2].UpdateTextboxText("2");
+        graph.AddNode(catchVariableNode, false);
+        graph.Connect(tryCatchNode.Outputs[1], catchVariableNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[1], catchVariableNode.Inputs[1], false);
+		graph.Connect(catchVariableNode.Outputs[0], returnNode.Inputs[0], false);
+
+		// Create the try block body
+		var parseNode = new MethodCall(graph);
+        parseNode.SetMethodTarget(new RealMethodInfo(nodeClass.TypeFactory, typeof(int).GetMethod("Parse", new[] { typeof(string) })!, nodeClass.TypeFactory.Get<int>()));
+        parseNode.Inputs[1].UpdateTextboxText("invalid");
+        graph.AddNode(parseNode, false);
+        graph.Connect(tryCatchNode.Outputs[0], parseNode.Inputs[0], false);
+
+		var tryVariableNode = new SetVariableValueNode(graph);
+		tryVariableNode.Inputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Variable
+		tryVariableNode.Inputs[2].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Value
+		tryVariableNode.Inputs[2].UpdateTextboxText("1");
+		graph.AddNode(tryVariableNode, false);
+
+		graph.Connect(parseNode.Outputs[0], tryVariableNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[1], tryVariableNode.Inputs[1], false);
+		graph.Connect(tryVariableNode.Outputs[0], returnNode.Inputs[0], false);
+
+		CreateStaticMainWithConversion(nodeClass, method);
+		
+		var output = Run<int>(project, options);
+
+        Assert.Equal(2, output);
+    }
+
+	[Theory]
+	[MemberData(nameof(GetBuildOptions))]
+	public void TestDeclareAndSetVariable(SerializableBuildOptions options)
+	{
+		var project = new Project(Guid.NewGuid());
+		var nodeClass = new NodeClass("Program", "Test", project);
+		project.Classes.Add(nodeClass);
+
+		var graph = new Graph();
+		var method = new NodeClassMethod(nodeClass, "MainInternal", nodeClass.TypeFactory.Get<int>(), graph, true);
+		nodeClass.Methods.Add(method);
+		graph.SelfMethod = method;
+
+		method.Parameters.Add(new("A", nodeClass.TypeFactory.Get<int>(), method));
+
+		var entryNode = new EntryNode(graph);
+		graph.AddNode(entryNode, false);
+
+		var declareVariableNode = new DeclareVariableNode(graph);
+		declareVariableNode.Inputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Initial value
+		declareVariableNode.Outputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Variable
+		graph.AddNode(declareVariableNode, false);
+
+		var setVariableValueNode = new SetVariableValueNode(graph);
+		setVariableValueNode.Inputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Variable
+		setVariableValueNode.Inputs[2].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true); // Value
+		graph.AddNode(setVariableValueNode, false);
+
+		var returnNode = new ReturnNode(graph);
+		graph.AddNode(returnNode, false);
+
+		graph.Connect(entryNode.Outputs[0], declareVariableNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[0], setVariableValueNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[1], setVariableValueNode.Inputs[1], false);
+		graph.Connect(entryNode.Outputs[1], setVariableValueNode.Inputs[2], false);
+		graph.Connect(setVariableValueNode.Outputs[0], returnNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[1], returnNode.Inputs[1], false);
+
+		CreateStaticMainWithConversion(nodeClass, method);
+
+		var output = Run<int>(graph.Project, options, [5]);
+
+		Assert.Equal(5, output);
+	}
+
+	[Theory]
+	[MemberData(nameof(GetBuildOptions))]
+	public void TestDeclareVariableDefaultValue(SerializableBuildOptions options)
+	{
+		var project = new Project(Guid.NewGuid());
+		var nodeClass = new NodeClass("Program", "Test", project);
+		project.Classes.Add(nodeClass);
+
+		var graph = new Graph();
+		var method = new NodeClassMethod(nodeClass, "MainInternal", nodeClass.TypeFactory.Get<int>(), graph, true);
+		nodeClass.Methods.Add(method);
+		graph.SelfMethod = method;
+
+		method.Parameters.Add(new("A", nodeClass.TypeFactory.Get<int>(), method));
+
+		var entryNode = new EntryNode(graph);
+		graph.AddNode(entryNode, false);
+
+		var declareVariableNode = new DeclareVariableNode(graph);
+		declareVariableNode.Inputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true);
+		declareVariableNode.Outputs[1].UpdateTypeAndTextboxVisibility(nodeClass.TypeFactory.Get<int>(), true);
+		graph.AddNode(declareVariableNode, false);
+
+		var returnNode = new ReturnNode(graph);
+		graph.AddNode(returnNode, false);
+
+		graph.Connect(entryNode.Outputs[0], declareVariableNode.Inputs[0], false);
+		graph.Connect(entryNode.Outputs[1], declareVariableNode.Inputs[1], false);
+		graph.Connect(declareVariableNode.Outputs[0], returnNode.Inputs[0], false);
+		graph.Connect(declareVariableNode.Outputs[1], returnNode.Inputs[1], false);
+
+		CreateStaticMainWithConversion(nodeClass, method);
+
+		var output = Run<int>(graph.Project, options, [5]);
+
+		Assert.Equal(5, output);
 	}
 }
